[16] pass validate tests

+++ include/util/str2t.hpp
@@ -11,7 +11,7 @@ int64_t intmax_to_int64(intmax_t intmax);
 uintmax_t string_to_uintmax(const std::string& s_uint, size_t line);
 uint32_t uintmax_to_uint32(uintmax_t uintmax);
 uint64_t uintmax_to_uint64(uintmax_t uintmax);
-void string_to_string_literal(const std::string& s_string, std::vector<int32_t>& string_literal);
+void string_to_string_literal(const std::string& s_string, std::vector<int8_t>& string_literal);
 int32_t string_to_char_ascii(const std::string& s_char);
 double string_to_double(const std::string& s_double, size_t line);
 uint64_t double_to_binary(double decimal);
